Create README.md
It is the modern implementation of the classical Flappy bird game. It is implemented in python with the help of popular module pygame.
In this game the bird is passing through the obstacles. Score updates after the bird passes through each obstacle. When bird collides with any obstacle or roof or floor, the game ends. The game also displays a high score and a current score.